+++ sandbox.rb
@@ -1,0 +1,3 @@
+pp ARGV
+ARGV.shift
+pp ARGV