+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 TODO
 
-- [ ] `punch plugin`
+- [ ] copy basics by request `punch basics`
+- [ ] copy samples by request `punch samples`
+- [ ] `punch new` should behave like `bundle gem`, if provided directory exists it just creates `punch.yml` and exec `bundle add`!
+- [ ] `punch plugin`, nothing special just regular SourceDecor
 - [ ] `punch extra`
 - [ ] create punch_users example, checking DSL, calculating status to measure SLOC for punched skeleton first and then complete domain with tests